c++/refactoring: move self_type for class registration to separate module
